# specs/002-bncc-explorer/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Planejador BNCC Catalog API
  version: 1.0.0
  description: Todas as rotas usam /api/v1 e exigem bearer token.
servers:
  - url: /api/v1
tags:
  - name: BNCC Catalog
  - name: BNCC Administration
paths:
  /bncc/levels:
    get:
      tags: [BNCC Catalog]
      summary: Lista níveis ativos
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Níveis ativos, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Level' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /bncc/levels/{levelId}/stages:
    get:
      tags: [BNCC Catalog]
      summary: Lista etapas ativas de um nível
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/LevelId' }]
      responses:
        '200': { description: Etapas, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Stage' } } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /bncc/skills:
    get:
      tags: [BNCC Catalog]
      summary: Consulta habilidades ativas com filtros, busca e paginação
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: levelId, in: query, schema: { type: string, format: uuid } }
        - { name: stageId, in: query, schema: { type: string, format: uuid } }
        - { name: year, in: query, schema: { type: integer, minimum: 1 } }
        - { name: axis, in: query, schema: { $ref: '#/components/schemas/Axis' } }
        - { name: q, in: query, schema: { type: string, maxLength: 100 }, description: Código completo/parcial ou texto }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: Página em código crescente; página além do total retorna data vazia e metadados, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedSkills' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/bncc/levels:
    get:
      tags: [BNCC Administration]
      summary: Lista níveis inclusive inativos (ADMIN)
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Níveis, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Level' } } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      tags: [BNCC Administration]
      summary: Cria nível (ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateLevel' } } } }
      responses: { '201': { description: Criado, content: { application/json: { schema: { $ref: '#/components/schemas/Level' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } }
  /admin/bncc/levels/{levelId}:
    patch:
      tags: [BNCC Administration]
      summary: Edita nome ou estado de nível (ADMIN)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/LevelId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateLevel' } } } }
      responses: { '200': { description: Atualizado, content: { application/json: { schema: { $ref: '#/components/schemas/Level' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /admin/bncc/levels/{levelId}/stages:
    get:
      tags: [BNCC Administration]
      summary: Lista etapas inclusive inativas de um nível (ADMIN)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/LevelId' }]
      responses: { '200': { description: Etapas, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Stage' } } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
    post:
      tags: [BNCC Administration]
      summary: Cria etapa vinculada ao nível (ADMIN)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/LevelId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateStage' } } } }
      responses: { '201': { description: Criada, content: { application/json: { schema: { $ref: '#/components/schemas/Stage' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } }
  /admin/bncc/stages/{stageId}:
    patch:
      tags: [BNCC Administration]
      summary: Edita etapa sem trocar seu nível (ADMIN)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/StageId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateStage' } } } }
      responses: { '200': { description: Atualizada, content: { application/json: { schema: { $ref: '#/components/schemas/Stage' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
  /admin/bncc/stages/{stageId}/skills:
    post:
      tags: [BNCC Administration]
      summary: Cria habilidade vinculada à etapa (ADMIN)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/StageId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateSkill' } } } }
      responses: { '201': { description: Criada, content: { application/json: { schema: { $ref: '#/components/schemas/Skill' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } }
  /admin/bncc/skills:
    get:
      tags: [BNCC Administration]
      summary: Consulta habilidades inclusive inativas (ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: levelId, in: query, schema: { type: string, format: uuid } }
        - { name: stageId, in: query, schema: { type: string, format: uuid } }
        - { name: active, in: query, schema: { type: boolean } }
        - { name: q, in: query, schema: { type: string, maxLength: 100 } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses: { '200': { description: Página administrativa, content: { application/json: { schema: { $ref: '#/components/schemas/PaginatedSkills' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
  /admin/bncc/skills/{skillId}:
    patch:
      tags: [BNCC Administration]
      summary: Edita habilidade sem alterar código ou etapa (ADMIN)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/SkillId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateSkill' } } } }
      responses: { '200': { description: Atualizada, content: { application/json: { schema: { $ref: '#/components/schemas/Skill' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }
  parameters:
    LevelId: { name: levelId, in: path, required: true, schema: { type: string, format: uuid } }
    StageId: { name: stageId, in: path, required: true, schema: { type: string, format: uuid } }
    SkillId: { name: skillId, in: path, required: true, schema: { type: string, format: uuid } }
  schemas:
    Axis: { type: string, enum: [PENSAMENTO_COMPUTACIONAL, MUNDO_DIGITAL, CULTURA_DIGITAL] }
    Level: { type: object, required: [id, code, name, active], properties: { id: { type: string, format: uuid }, code: { type: string }, name: { type: string }, active: { type: boolean } } }
    Stage: { type: object, required: [id, levelId, code, name, active], properties: { id: { type: string, format: uuid }, levelId: { type: string, format: uuid }, code: { type: string }, name: { type: string }, yearStart: { type: integer, nullable: true }, yearEnd: { type: integer, nullable: true }, active: { type: boolean } } }
    Skill: { type: object, required: [id, code, description, explanation, examples, active, stage], properties: { id: { type: string, format: uuid }, code: { type: string }, axis: { allOf: [{ $ref: '#/components/schemas/Axis' }], nullable: true }, description: { type: string }, explanation: { type: string }, examples: { type: array, minItems: 1, items: { type: string } }, active: { type: boolean }, stage: { $ref: '#/components/schemas/Stage' } } }
    PaginatedSkills: { type: object, required: [data, meta], properties: { data: { type: array, items: { $ref: '#/components/schemas/Skill' } }, meta: { type: object, required: [page, pageSize, total, totalPages], properties: { page: { type: integer }, pageSize: { type: integer }, total: { type: integer }, totalPages: { type: integer } } } } }
    CreateLevel: { type: object, required: [code, name], properties: { code: { type: string, enum: [EDUCACAO_INFANTIL, ENSINO_FUNDAMENTAL, ENSINO_MEDIO] }, name: { type: string, maxLength: 100 } } }
    UpdateLevel: { type: object, minProperties: 1, properties: { name: { type: string, maxLength: 100 }, active: { type: boolean } } }
    CreateStage: { type: object, required: [code, name], properties: { code: { type: string, maxLength: 40 }, name: { type: string, maxLength: 100 }, yearStart: { type: integer, nullable: true }, yearEnd: { type: integer, nullable: true } } }
    UpdateStage: { type: object, minProperties: 1, properties: { name: { type: string, maxLength: 100 }, yearStart: { type: integer, nullable: true }, yearEnd: { type: integer, nullable: true }, active: { type: boolean } } }
    CreateSkill: { type: object, required: [code, description, explanation, examples], properties: { code: { type: string, maxLength: 40 }, axis: { allOf: [{ $ref: '#/components/schemas/Axis' }], nullable: true }, description: { type: string, minLength: 1 }, explanation: { type: string, minLength: 1 }, examples: { type: array, minItems: 1, items: { type: string, minLength: 1 } } } }
    UpdateSkill: { type: object, minProperties: 1, properties: { axis: { allOf: [{ $ref: '#/components/schemas/Axis' }], nullable: true }, description: { type: string, minLength: 1 }, explanation: { type: string, minLength: 1 }, examples: { type: array, minItems: 1, items: { type: string, minLength: 1 } }, active: { type: boolean } } }
    Error: { type: object, required: [code, message], properties: { code: { type: string }, message: { type: string } } }
  responses:
    ValidationError: { description: Dados inválidos, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Sessão inválida, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Papel ADMIN necessário, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Recurso inexistente, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: Conflito de unicidade, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }